[TIMOB-23241] Logging macros for native module
