+++ templates/module/default/template/windows/src/{{moduleIdAsIdentifier}}.cpp.ejs
@@ -7,6 +7,7 @@
  */
 #include "<%- moduleIdAsIdentifier %>.hpp"
 #include "Titanium/detail/TiImpl.hpp"
+#include "TitaniumWindows/LogForwarder.hpp"
 
 using namespace Windows::ApplicationModel::Background;
